WIP: Group exceptions in palimpsest::exceptions
