+++ include/palimpsest/exceptions/TypeError.h
@@ -18,43 +18,32 @@
 
 #include <string>
 
-namespace palimpsest {
+#include "palimpsest/InternalError.h"
 
-//! Error with file and line references to the calling code.
-class InternalError : public std::runtime_error {
+namespace palimpsest::exceptions {
+
+//! Requested type doesn't match the one already in the dictionary.
+class TypeError : public InternalError {
  public:
-  /*! Create a new error.
+  /*! Create a type error.
    *
    * \param[in] file Source file of the instruction that threw the error.
    * \param[in] line Line of code in that file where the throw originates from.
    * \param[in] message Error message.
    */
-  InternalError(const std::string& file, unsigned line,
-                const std::string& message)
-      : std::runtime_error(message) {
-    std::ostringstream out;
-    out << "[" << file << ":" << line << "] " << message;
-    message_ = out.str();
-  }
+  TypeError(const std::string& file, unsigned line, const std::string& message)
+      : InternalError(file, line, message) {}
 
   /*! Copy an existing error, adding to the error message.
    *
    * \param[in] other Existing error.
    * \param[in] extra_message Additional error message.
    */
-  InternalError(const InternalError& other, const std::string& extra_message)
-      : std::runtime_error(other.message_ + extra_message),
-        message_(other.message_ + extra_message) {}
+  TypeError(const TypeError& other, const std::string& extra_message)
+      : InternalError(other, extra_message) {}
 
   //! Empty destructor
-  ~InternalError() throw() {}
-
-  //! Error message
-  const char* what() const throw() { return message_.c_str(); }
-
- private:
-  //! Complete error message
-  std::string message_;
+  ~TypeError() throw() {}
 };
 
-}  // namespace palimpsest
+}  // namespace palimpsest::exceptions
